+++ src/Interface/ITruckService.cs
@@ -5,7 +5,7 @@ namespace TruckGarage.Service;
 public interface ITruckService {
     Task<List<Truck>?> ListTrucksAsync();
     Task<Truck?> CreateTruckAsync(Truck truck);
-    Task<Truck?> FindTruckByIdAsync(int id);
-    Task<Truck?> UpdateTruckByIdAsync(int id, Truck truck);
-    Task<Truck?> RemoveTruckByIdAsync(int id);
+    Task<Truck?> FindTruckByIdAsync(long id);
+    Task<Truck?> UpdateTruckByIdAsync(long id, Truck truck);
+    Task<Truck?> RemoveTruckByIdAsync(long id);
 }
